ecu-simulator/pyproject.toml
[project]
name = "ecu-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-can>=4.6.1",
]
